MAINT: Move id resolving into its own module
